acpi: add a function to read a table and verify it
coreboot can currently read some ACPI tables from cbfs.
There is a fair amount of repeated code. Add a function
which can take a file name, and a table name, which
reads the file in, verifies its size, and its name,
returns a pointer to it if it is OK, NULL otherwise.

Show one use of the function: if a madt is found in
cbfs, use it instead of generating tables.

Over time, this code might replace other instances
of the copied-pasted code use for, e.g., the SLIT.

One such use of this code is on platforms where an
ACPI table has been supplied and it it not yet possible
to generate one.

Tested-by: Ronald G. Minnich <rminnich@google.com>

Change-Id: Id5e82eb0a82555b7763107e2a1e2a529a0c1b11f
Signed-off-by: Ronald G. Minnich <rminnich@google.com>