WIP: not for merge. Initial commit for rome psp board
The Rome PSP server boards release x86 from reset with
memory and most of the legacy chain (e.g. serial) ready to go.

This means the Rome is a lot like a qemu target. This work
explores the idea that we can dispense with AGESA and FSP
and UEFI entirely and put coreboot only in the x86 part
of SPI.

Change-Id: I4e5b0a28b0ad614949d61b4eb522a1166053105f
Signed-off-by: Ronald G. Minnich <rminnich@google.com>